moved function to data.util
