+++ pydgn/data/splitter.py
@@ -8,18 +8,8 @@ from torch_geometric.utils import negative_sampling, to_undirected, to_dense_adj
 
 import pydgn.data.dataset
 from pydgn.data.dataset import OGBGDatasetInterface
+from pydgn.data.util import to_lower_triangular
 from pydgn.experiment.util import s2c
-
-
-def to_lower_triangular(edge_index: torch.Tensor):
-    r"""
-    Transform Pytorch Geometric undirected edge index into its "lower triangular counterpart"
-    """
-    row, col = edge_index
-    lower_tri_mask = row > col
-    row, col = row[lower_tri_mask], col[lower_tri_mask]
-    lower_tri_edge_index = torch.cat((row.unsqueeze(0), col.unsqueeze(0)), dim=0)
-    return lower_tri_edge_index
 
 
 class Fold:
